requirements.txt
ccxt==4.0.103
PyYAML==6.0.1
pydantic==2.3.0
loguru==0.7.0